Fixing createFromBase
As it stands, the `createFromBase` static method used to recreate/recover an Illuminate `Request` object 
does not copy content across, despite trying to do so, when used outside of the scope of the class. Symfony's 
`$content` is a protected member, and as such, will be `NULL` in certain cases.

The `$content` member is accessible using a getter, and it makes more sense to use this in order to guarantee 
that the content will always be copied across.